requirements.txt
altair==5.4.1
antlr4-python3-runtime==4.9.3
anyio==4.6.0
appnope==0.1.4
astor==0.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.7
decorator==5.1.1
docker-pycreds==0.4.0
et-xmlfile==1.1.0
executing==2.1.0
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
imgaug==0.4.0
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.9.3
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.4
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyclipper==1.3.0.post5
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.0
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
samv2==0.0.4
scikit-image==0.24.0
scikit-learn==1.3.0
scipy==1.14.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
sentry-sdk==2.16.0
setproctitle==1.3.3
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.3
watchdog==5.0.3
wcwidth==0.2.13
